+++ packages/plugin-packages/mashroom-security-provider-openid-connect/src/security-provider/mashroom-bootstrap.ts
@@ -9,7 +9,6 @@ const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (plugin
         mode,
         issuerDiscoveryUrl,
         issuerMetadata,
-        rejectUnauthorized,
         scope,
         clientId,
         clientSecret,
@@ -19,16 +18,21 @@ const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (plugin
         usePKCE,
         rolesClaimName,
         adminRoles,
+        httpRequestRejectUnauthorized,
+        httpRequestTimeoutMs,
+        httpRequestRetry,
     } = pluginConfig;
 
     setClientConfiguration({
         issuerDiscoveryUrl,
         issuerMetadata,
-        rejectUnauthorized,
         clientId,
         clientSecret,
         redirectUrl,
         responseType,
+        httpRequestRejectUnauthorized,
+        httpRequestTimeoutMs,
+        httpRequestRetry,
     });
 
     setCallbackConfiguration({
@@ -37,7 +41,7 @@ const bootstrap: MashroomSecurityProviderPluginBootstrapFunction = async (plugin
         adminRoles,
     });
 
-    return new MashroomOpenIDConnectSecurityProvider(scope, usePKCE, extraAuthParams, rejectUnauthorized);
+    return new MashroomOpenIDConnectSecurityProvider(scope, usePKCE, extraAuthParams, httpRequestRejectUnauthorized);
 };
 
 export default bootstrap;